+++ admin/handlers/handlers.go
@@ -7,8 +7,8 @@ import (
 	"github.com/jmpsec/osctrl/carves"
 	"github.com/jmpsec/osctrl/logging"
 	"github.com/jmpsec/osctrl/metrics"
-	"github.com/jmpsec/osctrl/nodes"
 	"github.com/jmpsec/osctrl/pkg/environments"
+	"github.com/jmpsec/osctrl/pkg/nodes"
 	"github.com/jmpsec/osctrl/queries"
 	"github.com/jmpsec/osctrl/settings"
 	"github.com/jmpsec/osctrl/tags"
