+++ admin/handlers/types-templates.go
@@ -2,8 +2,8 @@ package handlers
 
 import (
 	"github.com/jmpsec/osctrl/carves"
-	"github.com/jmpsec/osctrl/nodes"
 	"github.com/jmpsec/osctrl/pkg/environments"
+	"github.com/jmpsec/osctrl/pkg/nodes"
 	"github.com/jmpsec/osctrl/queries"
 	"github.com/jmpsec/osctrl/settings"
 	"github.com/jmpsec/osctrl/tags"
