+++ logging/logging.go
@@ -2,7 +2,7 @@ package logging
 
 import (
 	"github.com/jmpsec/osctrl/backend"
-	"github.com/jmpsec/osctrl/nodes"
+	"github.com/jmpsec/osctrl/pkg/nodes"
 	"github.com/jmpsec/osctrl/queries"
 	"github.com/jmpsec/osctrl/settings"
 	"github.com/jmpsec/osctrl/types"
