+++ tls/handlers/utils.go
@@ -9,8 +9,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/jmpsec/osctrl/nodes"
 	"github.com/jmpsec/osctrl/pkg/environments"
+	"github.com/jmpsec/osctrl/pkg/nodes"
 	"github.com/jmpsec/osctrl/types"
 	"github.com/rs/zerolog/log"
 	"github.com/segmentio/ksuid"
